requirements.txt
gradio
ultralytics
opencv-python-headless
pandas
yt-dlp
pillow
